feat(studio): merge recap deck into Mission Control, sunset facts
Per review: one dashboard, not two. The agent-authored recap deck now renders
as a section under the Mission Control bento (RecapDeck, channel-coloured
cards). The separate /wrapped route + rail item are removed. The mechanical
'interesting facts' section is sunset entirely — it was mostly noise (top tool
= Read, raw tool-call counts, a tool mislabelled as a skill).

Verified via cmux: clock + bento + 12-card deck in one page, facts gone.
